Splitter deler av komponenten i to children for ansvarsfordeling
